Default operations

metagenomicsClientR.Default

Format

An R6Class generator object

Methods

AddNextflowEventNextflowEventsPost Add Nextflow Event Events from running Nextflow pipelines when using -with-weblog. See [the Nextflow documentation](https://www.nextflow.io/docs/latest/tracing.html#weblog-via-http). @param nextflow.event.model NextflowEventModel @returnType AnyType
status code : 200 | Successful Response return type : AnyType response headers :

status code : 422 | Validation Error return type : HTTPValidationError response headers :

AddUserUsersPost Add User @param user.model UserModel @returnType AnyType
status code : 200 | Successful Response return type : AnyType response headers :

status code : 422 | Validation Error return type : HTTPValidationError response headers :

DeleteUserUsersUidDelete Delete User @param uid integer @returnType AnyType
status code : 200 | Successful Response return type : AnyType response headers :

status code : 422 | Validation Error return type : HTTPValidationError response headers :

FilesChangeFilesChangesPost Files Change Track all changes in cloud storage @param body object @returnType AnyType
status code : 200 | Successful Response return type : AnyType response headers :

status code : 422 | Validation Error return type : HTTPValidationError response headers :

GenerateDownloadSignedUrlV4FilesSignedGet Generate Download Signed Url V4 Generates a v4 signed URL for downloading a blob. Note that this method requires a service account key file. You can not use this if you are using Application Default Credentials from Google Compute Engine or from the Google Cloud SDK. @param name character @returnType AnyType
status code : 200 | Successful Response return type : AnyType response headers :

status code : 422 | Validation Error return type : HTTPValidationError response headers :

GetNextflowEventNextflowEventsIdGet Get Nextflow Event @param id character @returnType AnyType
status code : 200 | Successful Response return type : AnyType response headers :

status code : 422 | Validation Error return type : HTTPValidationError response headers :

GetUserUsersUidGet Get User @param uid integer @returnType AnyType
status code : 200 | Successful Response return type : AnyType response headers :

status code : 422 | Validation Error return type : HTTPValidationError response headers :

ListFileEventsFilesChangesGet List File Events @param limit integer @param offset integer @param start.date character @param end.date character @param filename.regex character @returnType StorageEventCollection
status code : 200 | Successful Response return type : StorageEventCollection response headers :

status code : 422 | Validation Error return type : HTTPValidationError response headers :

ListNextflowEventsNextflowEventsGet List Nextflow Events @param limit integer @param offset integer @returnType AnyType
status code : 200 | Successful Response return type : AnyType response headers :

status code : 422 | Validation Error return type : HTTPValidationError response headers :

ListUsersUsersGet List Users @returnType AnyType
status code : 200 | Successful Response return type : AnyType response headers :

Public fields

apiClient

Handles the client-server communication.

Methods

Public methods

DefaultApi$new()

DefaultApi$AddNextflowEventNextflowEventsPost()

DefaultApi$AddNextflowEventNextflowEventsPostWithHttpInfo()

DefaultApi$AddUserUsersPost()

DefaultApi$AddUserUsersPostWithHttpInfo()

DefaultApi$DeleteUserUsersUidDelete()

DefaultApi$DeleteUserUsersUidDeleteWithHttpInfo()

DefaultApi$FilesChangeFilesChangesPost()

DefaultApi$FilesChangeFilesChangesPostWithHttpInfo()

DefaultApi$GenerateDownloadSignedUrlV4FilesSignedGet()

DefaultApi$GenerateDownloadSignedUrlV4FilesSignedGetWithHttpInfo()

DefaultApi$GetNextflowEventNextflowEventsIdGet()

DefaultApi$GetNextflowEventNextflowEventsIdGetWithHttpInfo()

DefaultApi$GetUserUsersUidGet()

DefaultApi$GetUserUsersUidGetWithHttpInfo()

DefaultApi$ListFileEventsFilesChangesGet()

DefaultApi$ListFileEventsFilesChangesGetWithHttpInfo()

DefaultApi$ListNextflowEventsNextflowEventsGet()

DefaultApi$ListNextflowEventsNextflowEventsGetWithHttpInfo()

DefaultApi$ListUsersUsersGet()

DefaultApi$ListUsersUsersGetWithHttpInfo()

DefaultApi$clone()

Method new()

Usage

DefaultApi$new(apiClient)

Method AddNextflowEventNextflowEventsPost()

Usage

DefaultApi$AddNextflowEventNextflowEventsPost(nextflow.event.model, ...)

Method AddNextflowEventNextflowEventsPostWithHttpInfo()

Usage

DefaultApi$AddNextflowEventNextflowEventsPostWithHttpInfo(
  nextflow.event.model,
  ...
)

Method AddUserUsersPost()

Usage

DefaultApi$AddUserUsersPost(user.model, ...)

Method AddUserUsersPostWithHttpInfo()

Usage

DefaultApi$AddUserUsersPostWithHttpInfo(user.model, ...)

Method DeleteUserUsersUidDelete()

Usage

DefaultApi$DeleteUserUsersUidDelete(uid, ...)

Method DeleteUserUsersUidDeleteWithHttpInfo()

Usage

DefaultApi$DeleteUserUsersUidDeleteWithHttpInfo(uid, ...)

Method FilesChangeFilesChangesPost()

Usage

DefaultApi$FilesChangeFilesChangesPost(body, ...)

Method FilesChangeFilesChangesPostWithHttpInfo()

Usage

DefaultApi$FilesChangeFilesChangesPostWithHttpInfo(body, ...)

Method GenerateDownloadSignedUrlV4FilesSignedGet()

Usage

DefaultApi$GenerateDownloadSignedUrlV4FilesSignedGet(name, ...)

Method GenerateDownloadSignedUrlV4FilesSignedGetWithHttpInfo()

Usage

DefaultApi$GenerateDownloadSignedUrlV4FilesSignedGetWithHttpInfo(name, ...)

Method GetNextflowEventNextflowEventsIdGet()

Usage

DefaultApi$GetNextflowEventNextflowEventsIdGet(id, ...)

Method GetNextflowEventNextflowEventsIdGetWithHttpInfo()

Usage

DefaultApi$GetNextflowEventNextflowEventsIdGetWithHttpInfo(id, ...)

Method GetUserUsersUidGet()

Usage

DefaultApi$GetUserUsersUidGet(uid, ...)

Method GetUserUsersUidGetWithHttpInfo()

Usage

DefaultApi$GetUserUsersUidGetWithHttpInfo(uid, ...)

Method ListFileEventsFilesChangesGet()

Usage

DefaultApi$ListFileEventsFilesChangesGet(
  limit = 100,
  offset = 0,
  start.date = NULL,
  end.date = NULL,
  filename.regex = NULL,
  ...
)

Method ListFileEventsFilesChangesGetWithHttpInfo()

Usage

DefaultApi$ListFileEventsFilesChangesGetWithHttpInfo(
  limit = 100,
  offset = 0,
  start.date = NULL,
  end.date = NULL,
  filename.regex = NULL,
  ...
)

Method ListNextflowEventsNextflowEventsGet()

Usage

DefaultApi$ListNextflowEventsNextflowEventsGet(limit = 100, offset = 0, ...)

Method ListNextflowEventsNextflowEventsGetWithHttpInfo()

Usage

DefaultApi$ListNextflowEventsNextflowEventsGetWithHttpInfo(
  limit = 100,
  offset = 0,
  ...
)

Method ListUsersUsersGet()

Usage

DefaultApi$ListUsersUsersGet(...)

Method ListUsersUsersGetWithHttpInfo()

Usage

DefaultApi$ListUsersUsersGetWithHttpInfo(...)

Method clone()

The objects of this class are cloneable with this method.

Usage

DefaultApi$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
####################  AddNextflowEventNextflowEventsPost  ####################

library(metagenomicsClientR)
var.nextflow.event.model <- NextflowEventModel$new() # NextflowEventModel | 

#Add Nextflow Event
api.instance <- DefaultApi$new()

result <- api.instance$AddNextflowEventNextflowEventsPost(var.nextflow.event.model)

####################  AddUserUsersPost  ####################

library(metagenomicsClientR)
var.user.model <- UserModel$new() # UserModel | 

#Add User
api.instance <- DefaultApi$new()

result <- api.instance$AddUserUsersPost(var.user.model)

####################  DeleteUserUsersUidDelete  ####################

library(metagenomicsClientR)
var.uid <- 56 # integer | 

#Delete User
api.instance <- DefaultApi$new()

result <- api.instance$DeleteUserUsersUidDelete(var.uid)

####################  FilesChangeFilesChangesPost  ####################

library(metagenomicsClientR)
var.body <- NULL # object | 

#Files Change
api.instance <- DefaultApi$new()

result <- api.instance$FilesChangeFilesChangesPost(var.body)

####################  GenerateDownloadSignedUrlV4FilesSignedGet  ####################

library(metagenomicsClientR)
var.name <- 'name_example' # character | 

#Generate Download Signed Url V4
api.instance <- DefaultApi$new()

result <- api.instance$GenerateDownloadSignedUrlV4FilesSignedGet(var.name)

####################  GetNextflowEventNextflowEventsIdGet  ####################

library(metagenomicsClientR)
var.id <- 'id_example' # character | 

#Get Nextflow Event
api.instance <- DefaultApi$new()

result <- api.instance$GetNextflowEventNextflowEventsIdGet(var.id)

####################  GetUserUsersUidGet  ####################

library(metagenomicsClientR)
var.uid <- 56 # integer | 

#Get User
api.instance <- DefaultApi$new()

result <- api.instance$GetUserUsersUidGet(var.uid)

####################  ListFileEventsFilesChangesGet  ####################

library(metagenomicsClientR)
var.limit <- 100 # integer | 
var.offset <- 0 # integer | 
var.start.date <- 'start.date_example' # character | 
var.end.date <- 'end.date_example' # character | 
var.filename.regex <- 'filename.regex_example' # character | 

#List File Events
api.instance <- DefaultApi$new()

result <- api.instance$ListFileEventsFilesChangesGet(limit=var.limit, offset=var.offset, start.date=var.start.date, end.date=var.end.date, filename.regex=var.filename.regex)

####################  ListNextflowEventsNextflowEventsGet  ####################

library(metagenomicsClientR)
var.limit <- 100 # integer | 
var.offset <- 0 # integer | 

#List Nextflow Events
api.instance <- DefaultApi$new()

result <- api.instance$ListNextflowEventsNextflowEventsGet(limit=var.limit, offset=var.offset)

####################  ListUsersUsersGet  ####################

library(metagenomicsClientR)

#List Users
api.instance <- DefaultApi$new()

result <- api.instance$ListUsersUsersGet()

}